+++ Tests/Integration/DSCResources/McAfee.config.ps1
@@ -30,8 +30,8 @@ configuration McAfee_config
         $Exception,
 
         [Parameter()]
-        [hashtable]
-        $BackwardCompatibilityException,
+        [string[]]
+        $ResourceParameters,
 
         [Parameter()]
         [object]