+++ Tests/Integration/DSCResources/Vsphere.integration.tests.ps1
@@ -7,6 +7,7 @@ $configFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCCompositeRe
 . $configFile
 
 $stigList = Get-StigVersionTable -CompositeResourceName $script:DSCCompositeResourceName
+$resourceParameters = (Get-DscResource -Module PowerSTIG -Name $script:DSCCompositeResourceName).Properties.Name
 
 $password = ConvertTo-SecureString -AsPlainText -Force -String 'ThisIsAPlaintextPassword'
 $credential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList 'Admin', $password