+++ Tests/Integration/DSCResources/WindowsFirewall.config.ps1
@@ -29,8 +29,8 @@ configuration WindowsFirewall_config
         $Exception,
 
         [Parameter()]
-        [hashtable]
-        $BackwardCompatibilityException,
+        [string[]]
+        $ResourceParameters,
 
         [Parameter()]
         [object]